+++ Test/ExperimantalFeaturesTest/InstallerTest.cs
@@ -1,6 +1,8 @@
 ﻿using ExperimentalFeatures;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
+using System;
 using System.IO;
+using System.Linq;
 using System.Threading.Tasks;
 
 namespace ExperimantalFeaturesTest
@@ -19,7 +21,7 @@ namespace ExperimantalFeaturesTest
         {
             _cachePath = Path.Combine(Path.GetTempPath(), "cache.json");
             var registry = new StaticRegistryKey();
-            var store = new DataStore(Constants.LogFile);
+            var store = new DataStore(new StaticRegistryKey(), Constants.LogFile);
             var feed = new LiveFeed(registry, Constants.LiveFeedUrl, _cachePath);
 
             _installer = new Installer(null, null, feed, store);
@@ -47,6 +49,38 @@ namespace ExperimantalFeaturesTest
             var hasUpdates = await _installer.CheckForUpdatesAsync();
 
             Assert.IsFalse(hasUpdates);
+        }
+
+        [TestMethod]
+        public async Task GetExtensionsMarkedForDeletionAsync()
+        {
+            string content = @"{
+            ""Add New File"": {
+                ""id"": ""2E78AA18-E864-4FBB-B8C8-6186FC865DB3"",
+                ""minVersion"": ""15.0"",
+                ""maxVersion"": ""15.2""
+                }
+            }";
+
+            File.WriteAllText(_cachePath, content);
+            await _installer.LiveFeed.ParseAsync();
+
+            var tooLow = _installer.GetExtensionsMarkedForDeletion(new Version(14, 0));
+            Assert.AreEqual(1, tooLow.Count());
+            Assert.AreEqual("2E78AA18-E864-4FBB-B8C8-6186FC865DB3", tooLow.ElementAt(0).Id);
+
+            var tooHigh = _installer.GetExtensionsMarkedForDeletion(new Version(16, 0));
+            Assert.AreEqual(1, tooHigh.Count());
+            Assert.AreEqual("2E78AA18-E864-4FBB-B8C8-6186FC865DB3", tooHigh.ElementAt(0).Id);
+
+            var lowerBounce = _installer.GetExtensionsMarkedForDeletion(new Version(15, 0));
+            Assert.IsFalse(lowerBounce.Any());
+
+            var upperBounce = _installer.GetExtensionsMarkedForDeletion(new Version(15, 2));
+            Assert.IsFalse(upperBounce.Any());
+
+            var middle = _installer.GetExtensionsMarkedForDeletion(new Version(15, 1));
+            Assert.IsFalse(middle.Any());
         }
     }
 }